python_backend/requirements.txt
flask
flask-cors
pinecone
groq
python-dotenv
gunicorn
